src/test/kotlin/co/csadev/advent2022/Day09Test.kt
/*
 * Copyright (c) 2022 by Charles Anderson
 */

package co.csadev.advent2022

import org.assertj.core.api.Assertions.assertThat
import org.junit.jupiter.api.Test

class Day09Test {

    private val input = """
        R 4
        U 4
        L 3
        D 1
        R 4
        D 1
        L 5
        R 2
    """.trimIndent().lines()

    private val input2 = """
        R 5
        U 8
        L 8
        D 3
        R 17
        D 10
        L 25
        U 20
    """.trimIndent().lines()

    private val example = Day09(input)
    private val example2 = Day09(input2)
    private val real = Day09()

    @Test
    fun `Matches example`() {
        val part1 = example.solvePart1()
        assertThat(part1).isEqualTo(13)

        val part2 = example2.solvePart2()
        assertThat(part2).isEqualTo(36)
    }

    @Test
    fun `Actual answer`() {
        val part1 = real.solvePart1()
        assertThat(part1).isEqualTo(6498)

        val part2 = real.solvePart2()
        assertThat(part2).isEqualTo(2531)
    }
}
